add black_list

+++ main.py
@@ -73,7 +73,7 @@ def getSetting(code: str, name=None):
             for arr in result.get('value'):
                 if arr['name'] == name:
                     return arr['value'] 
-        return result.get('value') 
+        return result.get('value')
 
 ADMIN_ARR = []
 for adm in list(getSetting('ADMINISTRATOR')):
@@ -845,6 +845,12 @@ def main_message(message):
     #write_json(message.json)
     logger.info('message.from_user.username: '+message.from_user.username)
     logger.info('message.text: ' + message.text)
+
+    black_list = getSetting('BLACK_LIST', message.from_user.username)
+    if black_list:
+        send_messages_big(message.chat.id, text=f'{message.from_user.username} заслужил пожиненный бан {black_list}', reply_markup=None)
+        send_message_to_admin(f'⚠️Внимание! \n {message.from_user.username} написал Джу:\n\n {message.text}')
+        return
 
     if isUserBan(message.from_user.username):
         bot.delete_message(message.chat.id, message.message_id)
